♻️🏢 Extract download trigger
